[IMP] account_report: hide line at 0
This commit allow to have a new options on reports that can hide line that are
equals to zero under certain conditions.

A line should be visible depending on its value and the ones of its children.
For parent lines, it's visible if there is at least one child with a value different from zero
or if a child is visible, indicating it's a parent line.
For leaf nodes, it's visible if the value is different from zero.

The option can have three state (never, by default, optional). Depending on the
report the state of the option will change.

Task: 3359936
Part-of: odoo/enterprise#45291