[FIX] planning: float error on shift end time upon applying template
Issue:
For some odd values of the template duration, for ex. 5h06, the
end time on the template will round to start time + duration but
round it at 5 min instead of 6.

Steps to reproduce:
- Install Planning
- Create a shift template from 8h with duration 5:06
- In planning, create a new shift and apply the template, observe
  the end time is rounded to 5 min instead of 6.

Cause:
For some values of the end time, when applying the `divmod`, the
result `mod` contains a floating point errors (0.0999999999964
instead of 0.1). When amplified by multiplying by 60 and then
converting literally with `int()`, this converts 5.999999999999979
to 5.

Fix:
Round before converting to `int`, just like what is done for the
start time.

Affected versions:
14.0 up to master

Reference:
opw-3418151

closes odoo/enterprise#44925

X-original-commit: ff4e39a09399f46790c0ad45bc9f9af4e4cd44d1
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>