[FIX] mrp_workorder: fix dialog calling dialog
Instead of creating a custom modal for the Employee Selection and Pin 
dialogs, we're using the preexisting Dialog modal.

This commit also fixes the modals which displayed a strange bar behind
the Employee Selection and Pin dialogs.
The cause of this was a Dialog being called within a Dialog 
component.

part of task-3231200

Part-of: odoo/enterprise#45317