[IMP] account: default taxes for multi-company product creation
Before:
When a new product was created in a multi-company setting, if the company field is empty, the product is available for all companies. The default sale and purchase tax of the company was set on the product. The problem is that only the default taxes of the currently active company was set on the product, so viewing the product in other companies showed an empty field for the tax.

Now:
When creating a new product with the company field empty, the default taxes of the other companies are set on the product as well.

Also, one query counter was updated as before the product used in the test had no taxes relevant to the company in env but since now all the default taxes are added during product creation, the product has one relevant tax, so we end up with seven more queries looking up tax related tables and adding the tax to the sale order.

task-3375286

closes odoo/enterprise#45075

Related: odoo/odoo#127196
Signed-off-by: William André (wan) <wan@odoo.com>