[FIX] account_online_synchronization: Fixing online_account_radio widget.
Since 8723f020, a new way to define items in radio_field.js is used.
The problem is that items are now defined asynchronous and it causes
that our widget tries to load balances data using these items when
we still don't have. It causes a traceback.

This fix load the balances data when the component is mounted and there
is 2 helpers functions that verifies if we have the data or not before
displaying it.

A test has been added to avoid this problem next time.

closes odoo/enterprise#44749

Task-id: none
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>