[IMP] web_grid,timesheet_grid: fix layout issues and utility classes
Instead of using custom CSS, we've switched to utility classes.
The hover effects are now handled with `--bg-opacity` CSS variables on
the absolutely-positioned elements.

Some of the hover effects have been adapted in the JS
(e.g.: the `.o_grid_row_timer` column no longer has a hover
effect when hovering another element in that column, but does have an
effect when hovering its row).

The `.o_grid_row_timer`'s column width defined in the grid-template
didn't correspond to the defined width in `$o-first-column-width`, creating an
unusual gap when fewer items were displayed.

The "zebra" background has been fixed using the `isEven` variable.
For this, an extra cell had to be added in the `o_grid_row_timer`
column.

The `TimesheetTimerHeader` has been placed above the grid instead of
inside it, thus allowing it to be full width of the screen without
taking the full width of the Month view when selected.

Other fixes include padding, text-truncation and
responsivity adjustments.

task-3340318
part of task-3326263

closes odoo/enterprise#45737

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>