[FIX] voip: call from mailing list contacts
when user clicks on call widget in mailing list contacts traceback will appear.

steps to produce:
1) install mass_mailing_sms and voip.
2) open one record and add mobile number.
3) click on mobile number.

Error: A traceback appears: "IndexError: list index out of range"

https://github.com/odoo/enterprise/blob/22992c1226268111ebf66f39214d8e15d0ede2e3/voip/models/voip_phonecall.py#L322
In the mailing.contact model, there is no field of type many2one with a
comodel name of res.partner. As a result, the partner_field_name returns an
empty list, and attempting to access the 0th element of an empty list is not
possible.

sentry-4306579710

closes odoo/enterprise#44807

X-original-commit: 37200401bdb1901a0abff895180945a657960576
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>
Signed-off-by: Bansi Ranpariya (bara) <bara@odoo.com>