[FIX] hr_payroll: analyze payslips without worked day
Issue:
------
In the report of payroll application (pivot view),
payslips (via its payslip lines) which contain no worked day
won't be taken into account for the amounts displayed.
It is therefore not possible to analyze them.

A concrete example are the payslips of the 13th month in the Belgian localization.

Cause:
------
To display these amounts, this SQL condition must be met:
```sql
SELECT
	...
	CASE WHEN wd.id = min_id.min_line THEN ...(display amount)... ELSE ... END
	...
FROM
	...
	left join hr_payslip_worked_days wd on (wd.payslip_id = p.id)
	...
	left join (select payslip_id, min(id) as min_line from hr_payslip_worked_days group by payslip_id) min_id on (min_id.payslip_id = p.id)
	...
```
Payslips that do not have a working day won't be listed in the
`hr_payslip_worked_days` table, so for these payslips we have:
`wd.id = NULL` and `min_id.min_line = NULL`.

In postgresql, the condition `NULL = NULL` is not true but unknown.

Solution:
---------
Modify the condition to handle the `NULL` value comparison like this:
`wd.id = min_id.min_line OR (wd.id IS NULL AND min_id.min_line IS NULL)`
This condition can be simplified as PostgreSQL supports `is not distinct from` operator:
`wd.id IS NOT DISTINCT FROM min_id.min_line`

opw-3429153

closes odoo/enterprise#44904

X-original-commit: 4dfeabb699fd7eda3dc8f9726430e5c08a8dd892
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>