[IMP] pos_preparation_display: remove top-level menu logic
The kitchen display icon in the main menu is opening directly
a preparation display under some conditions. This leads to problems
with some js tests (clickEverywhere for example) and is kinda not
Odoo-like as it does not follow the normal behaviour of every other
apps. It also does not provide direct access to the preparation display
settings when the conditions are met. This led to the change of
uniformizing this icon and making it so that the click on the app icon
will lead to a classical opening of the application (opening the
preparation displays main page).

closes odoo/enterprise#44198

Related: odoo/odoo#128699
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>