[FIX] mrp_workorder: shop floor fix continue production
Before this commit, when trying to use the 'continue production' button
on a register consumption quality check in the new shop floor display,
the user would get a traceback because the check would be updated first
and already point to the newly created next step that was not yet loaded
into the view.
This commit fixes this by not reloading the current quality check in the
view, as all the quality checks of this workorder will be reloaded
anyway.

Task: 3432610
Enterprise PR: https://github.com/odoo/enterprise/pull/44760

X-original-commit: ba149f7946222217839ae395690a3bfa87d9ac8f
Part-of: odoo/enterprise#45108