[FIX] web_studio: fix undeterministic report download test
Before this commit, the Studio test asserting the ReportEditor could download
the preview of the report with a certain context crashed undeterministically
because the overriden reoute in python returned with the wrong parameters, making the client
crash. Sometimes that crash did not occur because the browser was already closed.

After this commit, we do the assertions in JS, and we don't call the actual download
route anymore.

fixes runbot-error-23383

closes odoo/enterprise#44784

X-original-commit: 99b33d288ebb8c13cceb19b48403bf4355852442
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>