[FIX] web_gantt: avoid cell highlighting when dragging in Gantt
Before the conversion in Gantt, when a pill was dragged arount the Gantt
view, other pills were not highlighted when hovered. This behavior is
broken since the conversion. This commit fix this.

Steps
=====
- Install project(_enterprise)
- Open a project with multiple tasks with planned dates
set (e.g. `Research & Development` in the demo data
- Go to the Gantt view of the tasks of this project
- Click on a pill and drag it arount the Gantt view

Issue
=====
Other cell are higlighted when hovered

Cause
=====
The classes `o_grabbing`, `o_copying` and `o_no_dragging` are added to
the div with class `o_gantt_renderer` instead of `o_gantt_view`. The
selector used to define the conditions to provide visual feedback on
hover is therefore not valid anymore.

Fix
===
The structure of the stylesheet is modified to have the o_gantt_renderer
tag in the root selector as it is the one having the relevant class now
instead of o_gantt-view.

Note: It would have been WAY easier and clearer to simply add the
pseudo-class :has in the initial selector but, unfortunately, it is not
supported in Firefox.

task-3444256

closes odoo/enterprise#48505

Forward-port-of: odoo/enterprise#44737
X-original-commit: 1e24ead044b927e6a8c7386a911aac4e36b06184
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>