[FIX] mail_enterprise: add neutralize and improve VAPID keys generation
This commit adds the neutralize file to avoid sending WebPush
notification on "staging/duplicate/testing" databases.

This script cleans up the WebPush's keys and all related data in DB, so
no message is actually sent to the devices associated to the original
DB, but still let the "support/dev" use the feature in the regular
flow to subscribe to PUSH notification.

Also, we have modified how the VAPID keys are generated to be sure to
always have keys generated if they are deleted by the neutralize script.

Note:
Deleting the WebPush message in the CRON table aims to avoid sending
the wrong payload (wrong signature as the key as changed) to the
endpoints.

This can lead to weird behavior if many cron tasks are sending many
invalid payloads in batch. Which, from the endpoint point of view, can be
treated as an attempt to DDOS, Hacking, brut force, ...

Link:
https://www.odoo.com/documentation/master/administration/odoo_sh/getting_started/branches.html?#staging

closes odoo/enterprise#46164

X-original-commit: da614cb35be5995fd5cf9d77e4185cdda486908c
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>