[IMP] web_gantt: avoid default highlight of hovered pill at connect
Currently, when creating connector between pills, the default behavior
implemented in css is that hovered cells will be highlighted. This leads
to a lack of flexibility in managing which pill and what should be
displayed when hovered while creating connections.

This commit proposes to adopt the same behavior as the dragging/resizing
of pills, i.e. hovered pills are not highlighted by default, instead, the
highlighting is managed by the renderer.

task-3092632

Part-of: odoo/enterprise#44691