[IMP] project_enterprise: display cross project task dependencies
Currently, the dependencies between tasks is only displayed in the Gantt
view of a given project of the option 'Task dependency' is enabled on
that project. This commit extends the display of task dependencies to
the views "My Tasks" and "All Tasks".

Implemented functionalities:
- display the dependencies arrows and connectors in the 'my tasks' menu
- display a red fa-ban icon if the user tries to connect a task to
another one that belongs to a project on which the feature is disabled
- hide the connectors on the tasks that are in a project on which the
feature is disabled

task-3092632

closes odoo/enterprise#44691

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>