[IMP] account_reports: Prevent adding taxes on tax closing entry
This commit adds a constraint to the move lines to disallow users from
adding taxes when generating the tax closing entry.
The previous behavior allowed users to include taxes in
the closing entry, leading to functional issues and inconsistencies
in the tax report.
This fix ensures that move lines with associated taxes cannot be part of
a closing move.

closes odoo/enterprise#44770

Task-id: 3443259
Signed-off-by: Laurent Smet (las) <las@odoo.com>