[FIX] web_studio: reintroduce scss for icons in sidebar
This commit brings back the style of icons related to the fields
available in the sidebar. During the refactor of the report editor
(1), a lot of files present in the legacy folder were removed, and
this scss was not yet ported to current stylesheets.

With this fix, the style has been reintroduced without using a mixin,
since there is only one styling property involved. The visibility
of the icons has been preserved to the studio_sidebar.scss file.

task-3444791

1) https://github.com/odoo/enterprise/commit/2ae63d9a43ba973d905232105201d88e423a8245

closes odoo/enterprise#44923

X-original-commit: 94641a9cb511ec665b7109be1e86db46b2124d6b
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>