[FIX] hr_contract_salary: requested info on start
In salary configurator some advantages may have requested documents.
When the advantage is selected, the requested documents are shown and
required, while when the advantage is not selected, the requested
documents are hidden.

Problem:
On start, the requested documents of non-selected advantages are
still showing.

After this commit they will be hidden.

task-3444804

closes odoo/enterprise#44899

X-original-commit: 1e9764cd27f2130a3a8eded66fbee1bd97cb2d4f
Signed-off-by: Kevin Baptiste <kba@odoo.com>