[FIX] sign: modify selection field without loss of values
Current behaviour:
In sign module, when modifying a Selection type field in a template:
- if you add an option value, exit and re-open the template, the only remaining value is
the last one you added
- if you remove an option value, exit and re-open the template, there is no more value.

Expected behaviour:
Get the right values set in the Selection field

Steps to reprpoduce:
- Go to Sign, upload a pdf template.
- Add a selection field with 2 values (eg. "a" and "b") and save.
- Exit and re-open the modified template(everything is ok).
- Modify the field added, remove or add a value(eg. "c").
- Validate, exit and re-open the template.

Issue:
There is no value or only one("c") in the selection field.

Reason of the problem:
When updating the selection field in ``SignTemplate.update_from_pdfviewer()``,
the dictionary used to update the ``SignItem`` has the field ``option_ids`` for values as a list
of strings and those strings are meant to be ``SignItemOption.id``. So when ``SignItem``
overwrites its values, invalid ``SignOptionItem`` objects are returned but no error is raised.

opw-3201124

closes odoo/enterprise#44855

Signed-off-by: Eteil Junior Djoumatchoua (etdj) <etdj@odoo.com>