[FIX] l10n_de_pos_cert: check taxes when importing sale order
This PR combines the fixes in https://github.com/odoo/enterprise/pull/40997/,
https://github.com/odoo/enterprise/pull/44375
and https://github.com/odoo/enterprise/pull/44483 for version 16 and
upward.

Steps to reproduce:

1. make a sale order for a product with no taxes
2. open a PoS session, click on Quotation / Order
3. select the sale order, and validate
4. attempt to close the session

This will lead in an error in Fiskaly, that blocks closing of the
session. The taxes of order lines should be checked before they are
imported in the PoS.

The original fix (https://github.com/odoo/enterprise/pull/40997)
did not consider `add_orderline` method was also used to
display old orders in the PoS, so the forward ports were not merged in
version 16 up to master (with the exception of saas-16.2).

The other fix (https://github.com/odoo/enterprise/pull/44375) made sure
that only new orders are checked for taxes. In turn PR https://github.com/odoo/enterprise/pull/44483
made the order state available for the check (see PR description)

This PR will add the fix for that issue (https://github.com/odoo/enterprise/pull/44483)
to address both bugs simultaneously in version 16 and later.

opw-3430698

closes odoo/enterprise#45627

X-original-commit: ac2e9b868f4920e20d68827d9c8e22e37f223de9
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>