[FIX] l10n_mx_edi: Fix CFDI with Exento taxes
Such taxes must not be ignored but some attributes should be invisible.
For example, the base of a 0% tax shouldn't be there.

However, when dealing with a 0% tax being not Exento, the tax amounts must be there even
if equal to zero.

Based on feedbacks received from Mexican users to make the CFDI documents passing the SAT validation.

closes odoo/enterprise#45106

X-original-commit: 1fe2232d4e41fec1b04702cb855ae73965251d22
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>