[FIX] account_{accountant,reports}: fix deferrals in future periods
Currently the way deferrals work, we expect the deferral to always start
in the period of the move date.

For the `at_validation` method, this caused the issue that no entry was
generated for a single future period, since we assumed it was the same
month as the move date and thus no deferral was needed.

For the `manual` method, we could not create entries for a future single
period, and also no intermediate entries to carry over the cost from the
move period to the start period of the deferral.

This commit fixes all the above issues and allows to both create
deferrals for a single period via both methods, and allows to
correctly defer the amounts from the move period to the start of the
deferral period for the `manual` method.

task-3441429

closes odoo/enterprise#46047

X-original-commit: 519bf6e32d8070e4b58cce122ad3a60a6667e102
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>