[REF] *: tests: remove legacyExtraNextTick
Now that most of the webclient codebase (fields, views, client
actions...) has been converted to owl, the legacy extra next tick
used in a lot of tests is no longer necessary. This commit removes
the helper and its usage. At some places, a real tick was needed,
but we didn't see it because of the use of the legacy extra next
tick.

Part of task~3439226

closes odoo/enterprise#44869

Related: odoo/odoo#130236
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>