[FIX] account_reports: prevent RangeError on Chrome
For large reports (>20000 lines Aged Payable in the customer DB), Chrome
may throw a RangeError: Maximum call stack size exceeded. This is
because the call to this.$el.find('tr') is not restrictive enough,
thus raising a RangeError in jQuery's find function.

This commit changes the find's selector to only return rows with
the parent-id data attribute set. We can do that because rows
without parent-id won't be processed in the next map call. So there
is no point in returning them.

NB: it can probably also happen in Firefox/Other browsers but with
bigger reports.

opw-3275072

closes odoo/enterprise#44883

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>