[IMP] web_grid: avoid using parseFieldNode in grid arch parser
Before this commit, `parseFieldNode` from Field component was used
to get label for measure field, to get widget and invisible values.

This commit avoids using it and search the attributes needed directly
on the node because `parseFieldNode` also search the field component
to use but in the grid view it is not needed since the field registry
is no longer used for section and row fields.

task-3222154

Part-of: odoo/enterprise#44599