[FIX] web_gantt: show pill dependencies in Firefox
The Gantt dependencies display in the Gantt view is not working anymore
in Firefox browser since the conversion to owl. This commit fixes that.

Steps
=====
Using firefox as a browser:
- Install module project
- Activate task dependencies in the settings of Project
- Open the Gantt view of a project with task dependencies
enabled (e.g. Research and Development).

Issue
=====
Dependencies between tasks are not displayed in the Gantt view.

Cause
=====
The `d` property of the tag `path` should contained data separated by
spaces. Currently, commas are added to separate those datas. This is not
supported by Firefox and therefore leads to the arrows not to be
displayed.

Fix
===
No more commas are added to separate geometrical datas during the
generation of dependency arrows.

task-3447057

closes odoo/enterprise#45641

X-original-commit: 1425d8eb544f417938b827bc51947230b811bf8c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>