[FIX] web_grid: fallback on count measure if no measure field
Before this commit, when a grid view is defined without any measure
field in its definition, a traceback is occured because the "__count"
field is not inside the fieldsInfo which is expected.

This commit reviews the measureField attribute in archInfo to
determine if the field is readonly or not and then improves
editable attribute in archInfo to determine if the grid view
can be editable or not including the readonly attribute of
measure field. By doing that, it is no longer needed to check
the field inside fields info in the grid controller.

task-3222154

closes odoo/enterprise#44599

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>