[IMP] web_studio: improve approval rules
This commit improves the approvals feature by adding multiple
behaviors.

Firstly, the UI of the form and list views corresponding to the
approval rules has been revamped and improved with new tooltips,
new options and better labels.

Those new options are the following:

- users_to_notify: list of users to notify by a note in the chatter by
mentionning users in the message body.

- notification_order: when approvals have a higher notification_order,
activities and message to ask for approval is postponed until all
approval entries from  rules with lower notification order  are approved.

Secondly, the notification can now display a message that is editable,
instead of a generic list of missing approvals.

The name of the rule is no longer computed, but is assigned when created
from Studio, using the button label and the associated model.

Finally, the Studio sidebar has been updated with the newest additions,
that are fully editable. The code has been reworked to use the Record
component, as well as fields components.

Approvals tests have been adapted to match and test the newest features.
A new python test verify that notification_order is taken in count, in
order to request new approvals.

task-3444754

closes odoo/enterprise#46446

Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>