[FIX] account_followup: download document with date in filename
Issue:
======
When you download a document with a date in it's name and the date
has dots  (.) it will mess the with file at it expects the type of the
file after the first dot.

Steps to reproduce the error :
==============================
1-Install accounting
2-Install German Language
3-Go to accounting/customer/followup reports
4-Select an item and do follow up
5-Print the report.
6-You will get a report not in the pdf format.

Solution:
=========
I replaced the '.' with '-' in the date formats of the name.

opw-3425241

closes odoo/enterprise#44932

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>