[FIX] account_followup: download document with date in filename
Issue:
======
When you download a document with a date in it's name and the date
has dots  (.) it will mess the with file at it expects the type of the
file after the first dot.

Steps to reproduce the error :
==============================
1-Install accounting
2-Install German Language
3-Go to accounting/customer/followup reports
4-Select an item and do follow up
5-Print the report.
6-You will get a report not in the pdf format.

Solution:
=========
I replaced the '.' with '-' in the date formats of the name.

opw-3425241

closes odoo/enterprise#45073

X-original-commit: 89d4f05efab9e0a170a2d9b92b77d9199db72466
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>