[FIX] account_reports: make deferral banner clickable
Currently when you open the deferred expense or revenue report and there
already exist deferral moves for that period, we show a banner
indicating that. However the user has no direct way of going to the
involved entries. This commit makes the message clickable to that a
user can quicly access the mentioned entries.

Also the message mentions the entries are posted, but they are not
always. So we renamed the term to 'generated' instead.

Furthermore an issue is fixed where you couldn't generate a grouped
deferral entry when there already was an automatic one on the same date
from before (when the mode was automatic).

task-3441804

closes odoo/enterprise#46844

X-original-commit: c5aedd2eff1d7110ece0d0a4f0d9d37cc941db75
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>