[FIX] planning: copy occurrences' date to last day of the month
Steps:
In planning, create a shift with date: 31/1/2020
and recurrence each month for 5 months.

Expected:
For all shifts, the date should be the last day of the month.

Issue:
From february on, each shift's date is the 29th.

Cause:
We calculate the next date by adding the recurrence delta to the
previous one. In this case:
31/1/2020 + 1 month = 29/2/2020; 29/2/2020 + 1 month = 29/3/2020; etc.

Fix:
Calculate the nth next date by adding the recurrence delta,
multiplied by n, to the first date. In this case:
31/1/2020 + 1 month = 29/2/2020; 31/1/2020 + 2 month = 31/3/2020; etc.

task-3449851

X-original-commit: e847c18fd35e304929029769409b3670473ec2f7
Part-of: odoo/enterprise#50613