[IMP] l10n_mx_edi: simplify company rule on the certificate
This commit adds a required `company_id` on `l10n_mx_edi.certificate`
and changes the `l10n_mx_edi_certificate_ids` on `res.company` to a One2many field.

This allows simplifying the security rule.
Also, instead of calling a subprocess to run an openssl command to convert a key from
der format to pem, this commit makes use of existing `load_der_private_key` method to
load the key, so that it's then possible to get it in the pem format. It is also now
more similar how it is done in other l10n modules where certificates are loaded.

task-3276992

closes odoo/enterprise#44971

Related: odoo/upgrade#5011
Signed-off-by: Laurent Smet (las) <las@odoo.com>