[FIX] mrp_workorder: Prevent duplicate quality checks creation
Extract `processed_move` outside the loop of workorder.
This is necessary to track moves that have been registered in precedent workorders so that it does not generate new checks in moves_without_check.

This change is a bit hacky, however moves_without_check was removed in master (saas-16.4), so it will not have an impact on future versions.

OPW-3322942

closes odoo/enterprise#45000

X-original-commit: be0059116226d0262c9573c0b4567489ed67ea75
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: David Fesquet (dafr) <dafr@odoo.com>