[FIX] knowledge: reset the scroll offset when selecting a template
When a user switches from one template to another in the template
gallery of Knowledge, the scroll view of the template preview does
not reset, causing inconvenience as the user has to manually scroll
back to the top to start read the new selected template. This commit
aims to resolve this issue by automatically resetting the scroll
offset when selecting a new template, providing a smoother user
experience.

Steps to reproduce the issue:
1. Open the template gallery of Knowledge
2. Select a template to preview it
3. Scroll in the template preview a bit
4. Select another template to preview it
=> The scroll offset of the scroll view is not reset.

TO BE: When the user selects another article, the scroll offset
should be set to 0 (back to the top).

task-3508381

closes odoo/enterprise#47627

X-original-commit: 41b7508bf5ecf4757eaea2488b1dbe5193a7d101
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>