[FIX] stock_barcode_mrp: removing duplicate scrap button
Issue:
-------------------
In the barcode operations settings, the scrap button is duplicated

Steps to reproduce:
-------------------
* Go to barcode -> Operations
* In Delivery orders -> To process -> Click on some record
* Click on the settings icon in the top right menu
* There are two scrap buttons

Cause:
-------------------
When we have the barcode and MRP apps both installed, the scrap button is loaded twice, one from the javascript function printButtons() and the other from the file main.xml from stock_barcode_mrp add-on.  The commit 61ca0bd fixed the javascript file and now there is no need to keep the second scrap button.

OPW-3440927

closes odoo/enterprise#45615

X-original-commit: e5da5e72b2858d2669d73c579cd8fd073881c206
Signed-off-by: Tiffany Chang <tic@odoo.com>