[FIX] spreadsheet_edition: omit graph order & measure from context
When you apply a saved filter (with order) on graph and then change its
measure or order, the new changes are discarded and data is inserted as
per filter in spreadsheet since order and measure of context take
precedence over applied changes stored in metadata.

This commit omits graph order and measure from context while inserting
the graph to ensure that latest changes are applied.

closes odoo/enterprise#45017

Task: 3373768
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>