[FIX] mrp_workorder: shop floor show quality check title in popup
Before this commit, when opening a quality check in the new shop floor
display, the title of the quality check was not shown. This could lead
to confusion in the case of register component quality checks, where it
would not be clear which component we are registering.
This commit fixes this by adding the title of the quality check to the
top of the quality check instructions, as well as adding the component
name in case of a record_ type step.

Task: 3432610
Enterprise PR: https://github.com/odoo/enterprise/pull/44760

X-original-commit: cbfca4f391083a75af5df8284c44afe262a63382
Part-of: odoo/enterprise#45108