[FIX] pos_sale_stock_renting : override compute qty delivered
For POS rental order products the qty_delivered should not be aggregated as they get returned. When a product was returned in the rental order, the delivered qty would increase and this would prevent the qty_returned and qty_delivered to be equal if the original value was 1.

Steps to Reproduce on Runbot:

1. Create a product in the rental app with default settings (able to be rented)

2. Create a new rental order with any customer and quantity = 1 of this product.

3. Save the order but DON'T confirm it.

4. Open new POS session. (default shop works fine)

5. Click the quotation/order button, select the created rental order, settle the order, and click yes if it says to import product.

6. Click the payment button, and validate with Bank payment.

7. Go back and refresh the rental order in the rentals app (state should be "Picked-up")

8. Click the "RETURN" button and validate the single item returned.

Current Behavior:
The delivered qty would increase by 1.

Expected Behavior:
The delivered qty should not change as we are only returning the product.

OPW-3393540

closes odoo/enterprise#45023

Signed-off-by: Robin Heinz (rhe) <rhe@odoo.com>