[FIX] account_consolidation: don't show account_consolidation's hierarchy option on reports not supporting it
To reproduce:

1) Install account_reports and account_consolidation
2) Open the generic Balance Sheet
3) Click on "Options" filter
==> Hierachy and Subtotals is displayed twice, and only the first one is working.

The second occurrence is actually the 'consolidation_hierarchy' option. It is displayed because the extra options template is overridden in account_consolidation by directly modifying the parent template (because there is no field for overriding this template with a new one on account.report), and it lacks a t-if. So, other reports not supporting this option all display it.

closes odoo/enterprise#45025

Signed-off-by: John Laterre (jol) <jol@odoo.com>