[FIX] mrp_workorder: shop floor manufacturing done no final product lot
Before this commit, when marking a MO for a tracked product as done in
the new shop floor display, but with no lot/serial set, the user would
only get an error after the 10-second validation timeout.
This commit fixes this by checking for this case in pre_button_mark_done
and immediately raising a UserException in this case.

Task: 3432610
Enterprise PR: https://github.com/odoo/enterprise/pull/44760

X-original-commit: 42222ac054bf867ce74ea5c2f06d66545c2e5d6e
Part-of: odoo/enterprise#45108