[FIX] mrp_workorder: shop floor ignore qty_producing when running check
Before this commit, when trying to execute a quality check on a WO with
qty_producing set to zero in the new shop floor display, it would show
a UserError.
This commit fixes this by ignoring the value of qty_producing when
running a quality check.

Task: 3432610
Enterprise PR: https://github.com/odoo/enterprise/pull/44760

X-original-commit: 9066e38e2b71a6e65eca79e3b2e45fa2a38b9d16
Part-of: odoo/enterprise#45108