[FIX] sale_subscription : assign renewed subscription end_date
Issue:
======
When creating a renewal subscription the renewed subscription 'end_date'
is assigned to today and not the start date of the renewal.

Steps to reproduce the error:
=============================
- Create a monthly subscription (select any customer and monthly
subscription plan for easy setup)
- Confirm the subscription , create invoice and confirm it
- Renew the subscription order and select recurrence yearly and confirm
  it
- Go back to the renewed subscription :  it is locked and the end date
  is equal to today

Solution:
=========
I moved the 'end_date' update after the 'set_close' , because in 'set_close'
it updates the end_date as today.

opw-3420933

closes odoo/enterprise#45700

X-original-commit: e2ed421f1e0d0b81d2413728272df0a7fb216a7a
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>