[FIX] stock_barcode_mrp: scan several kits
To reproduce the issue:
1. Create two kits with defined barcode
2. Barcode > Operations > Receipt, Create
3. Scan both kits
4. Validate

Error: an Odoo Servor occurs "TypeError: StockMove.action_explode()
takes 1 positional argument but 2 were given"

When calling the ORM, we don't provide the SMs' IDs in an array. As
a result:
https://github.com/odoo/odoo/blob/e902713648bca329e6821859462b2128aad62c09/odoo/api.py#L447
Only the first ID will be stored in `ids`. The next IDs will be
considered as method arguments.

closes odoo/enterprise#45642

X-original-commit: e98dc378a55d371a823f89fc3c41b387dcc0621d
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>