[IMP] knowledge: cover portal user use case with a flow tour
This commit adds a "main flow" testing tour for the portal user.

Testing this flow is important as it requires initializing a web-client for
portal users, mainly to allow them to write on articles, which can be
technically tricky.

This commit contains a simplified version of the tour for internal users, with
specific steps to cover certain features and use cases for the portal, mainly:
- Can only access articles to which he has specific access
- Allowed to create private articles
- Allowed to create children articles under "workspace" articles to which he
  has direct write access to
- Can mark articles as favorite and re-order his favorites
- Can use a specific search feature to find articles

Task-3434132

X-original-commit: 2b81c770d60e0ab5487c07877d104aa4f1e69038
Part-of: odoo/enterprise#45968