[IMP] *helpdesk*: improve help center
This commit introduces various improvements in the Help Center on the
website.

Banner Edition
==============
Purpose: allow users to customize the help webpage, as it is possible
for other web pages

Currently the banner of the help center is not editable by the user as
it contains a search field. The search field is moved out of the cover
to allow the the user to personalize it.

Publish Button
==============
Purpose: move the published button to the systray to be coherent with
what is usually done in website + we don't need it anymore now that we
have 1 help web page per helpdesk team.

- help webpage: move the published button to the systray
- ticket submission webpage: move the published button to the systray

Knowledge Article
=================
Purpose: allow users to easily edit the article set on the team from
there.

- helpdesk.team form view > knowledge feature > display the external
link icon next to the article field

task-3186564

closes odoo/enterprise#45115

Related: odoo/odoo#134976
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>