[FIX] account_avatax: always send positive quantities
Avatax returned positive tax amounts for records with a single line
with a negative quantity.

The Avalara documentation specifies that the quantity field should
always be positive, even when it's a return or refund [1]. The sign is
determined by 'amount'.

Changing the larger, overarching test_*_odoo_invoice tests to include
negative quantities was considered but doing so doesn't catch this
bug. Those tests ensure Odoo does the right thing with the Avatax
responses. That works correctly. Instead the issue lies in the request
we send to Avatax. A simple test was added to ensure only positive
quantities are sent.

[1] "This quantity value should always be a positive value
representing the quantity of product that changed hands, even when
handling returns or refunds."
https://developer.avalara.com/api-reference/avatax/rest/v2/methods/Transactions/CreateTransaction/

PS. Before this patch another incorrect behavior occurred on
sale.order. The amounts on a line with a negative quantity would flip
between positive and negative after each button_update_avatax(). This
happened because account_avatax_sale sets price_subtotal on the line
from the Avatax response. Since we send price_subtotal as 'amount' to
Avatax it would alternate between positive and negative because we
also sent the negative quantity. It's not necessary to set
price_subtotal (US only has price-excluded taxes and we don't do it on
account.move), but because it doesn't cause any immediate issues it
wasn't changed.

opw-3431324

closes odoo/enterprise#47435

X-original-commit: e29765bad32344d45a70c6d22fb350a4e0103e34
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>